Update declaration parsing and support block name selectors.
Some types have been modified to make the semantics of the parsers
clearer, the declaration parser is now abstracted over a parser for
identifiers (to support normal identifiers, selector syntax, and
union tags), and blocks now optionally contain an additional
selector that enforces that the block be named. By default this
has type `ident`, but a different type can be specified. It likely
only makes sense for this name to have type ident, string, or int -
anything else would be too structured.